+++ moochat/db/services.php
@@ -4,12 +4,11 @@
 // Moodle is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
+// any later version.
 /**
- * Define all the restore steps that will be used by the restore_moochat_activity_task
+ * External services definition for block_moochat
  *
- * @package    mod_moochat
+ * @package    block_moochat
  * @copyright  2025 Brian A. Pool
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -17,13 +16,20 @@
 defined('MOODLE_INTERNAL') || die();
 
 $functions = [
-    'mod_moochat_send_message' => [
-        'classname'   => 'mod_moochat\external\send_message',
+    'block_moochat_send_message' => [
+        'classname'   => 'block_moochat\external\send_message',
         'methodname'  => 'execute',
-        'description' => 'Send a chat message and receive AI response',
+        'description' => 'Send a message to the AI chatbot',
         'type'        => 'write',
         'ajax'        => true,
-        'capabilities'=> 'mod/moochat:submit',
+        'loginrequired' => true,
+    ],
+    'block_moochat_save_conversation' => [
+        'classname'   => 'block_moochat\external\save_conversation',
+        'methodname'  => 'execute',
+        'description' => 'Save conversation messages to database',
+        'type'        => 'write',
+        'ajax'        => true,
         'loginrequired' => true,
     ],
 ];