+++ moochat/db/upgrade.php
@@ -4,12 +4,11 @@
 // Moodle is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
+// any later version.
 /**
- * Define all the restore steps that will be used by the restore_moochat_activity_task
+ * Upgrade script for block_moochat
  *
- * @package    mod_moochat
+ * @package    block_moochat
  * @copyright  2025 Brian A. Pool
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -17,55 +16,43 @@
 defined('MOODLE_INTERNAL') || die();
 
 /**
- * Upgrade code for mod_moochat
+ * Upgrade code for the moochat block
  *
- * @param int $oldversion the version we are upgrading from
- * @return bool always true
+ * @param int $oldversion
+ * @return bool
  */
-function xmldb_moochat_upgrade($oldversion) {
+function xmldb_block_moochat_upgrade($oldversion) {
     global $DB;
-
     $dbman = $DB->get_manager();
 
-    // Add display field
-    if ($oldversion < 2025103002) {
-        $table = new xmldb_table('moochat');
-        $field = new xmldb_field('display', XMLDB_TYPE_INTEGER, '2', null, XMLDB_NOTNULL, null, '0', 'introformat');
-        
-        if (!$dbman->field_exists($table, $field)) {
-            $dbman->add_field($table, $field);
-        }
-        
-        upgrade_mod_savepoint(true, 2025103002, 'moochat');
-    }
+    if ($oldversion < 2026021100) {
+        // Define table block_moochat_conversations to be created.
+        $table = new xmldb_table('block_moochat_conversations');
 
-    // Add include_section_content and include_hidden_content fields
-    if ($oldversion < 2025103003) {
-        $table = new xmldb_table('moochat');
-        
-        $field1 = new xmldb_field('include_section_content', XMLDB_TYPE_INTEGER, '1', null, XMLDB_NOTNULL, null, '0', 'display');
-        if (!$dbman->field_exists($table, $field1)) {
-            $dbman->add_field($table, $field1);
-        }
-        
-        $field2 = new xmldb_field('include_hidden_content', XMLDB_TYPE_INTEGER, '1', null, XMLDB_NOTNULL, null, '0', 'include_section_content');
-        if (!$dbman->field_exists($table, $field2)) {
-            $dbman->add_field($table, $field2);
-        }
-        
-        upgrade_mod_savepoint(true, 2025103003, 'moochat');
-    }
+        // Adding fields to table block_moochat_conversations.
+        $table->add_field('id', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, XMLDB_SEQUENCE, null);
+        $table->add_field('instanceid', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, null);
+        $table->add_field('userid', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, null);
+        $table->add_field('role', XMLDB_TYPE_CHAR, '20', null, XMLDB_NOTNULL, null, null);
+        $table->add_field('message', XMLDB_TYPE_TEXT, null, null, XMLDB_NOTNULL, null, null);
+        $table->add_field('timecreated', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, null);
 
-    // Add chatsize field
-    if ($oldversion < 2025103004) {
-        $table = new xmldb_table('moochat');
-        $field = new xmldb_field('chatsize', XMLDB_TYPE_CHAR, '10', null, XMLDB_NOTNULL, null, 'medium', 'include_hidden_content');
-        
-        if (!$dbman->field_exists($table, $field)) {
-            $dbman->add_field($table, $field);
+        // Adding keys to table block_moochat_conversations.
+        $table->add_key('primary', XMLDB_KEY_PRIMARY, ['id']);
+        $table->add_key('instanceid', XMLDB_KEY_FOREIGN, ['instanceid'], 'block_instances', ['id']);
+        $table->add_key('userid', XMLDB_KEY_FOREIGN, ['userid'], 'user', ['id']);
+
+        // Adding indexes to table block_moochat_conversations.
+        $table->add_index('instanceid-userid', XMLDB_INDEX_NOTUNIQUE, ['instanceid', 'userid']);
+        $table->add_index('timecreated', XMLDB_INDEX_NOTUNIQUE, ['timecreated']);
+
+        // Conditionally launch create table for block_moochat_conversations.
+        if (!$dbman->table_exists($table)) {
+            $dbman->create_table($table);
         }
-        
-        upgrade_mod_savepoint(true, 2025103004, 'moochat');
+
+        // Moochat savepoint reached.
+        upgrade_block_savepoint(true, 2026021100, 'moochat');
     }
 
     return true;